demo of verifying a MIP.
This completes the critical path of using tessera to produce
transparency logs, and creating and verifying MIPs for the log's
entries, and using the signed note and checkpoint specs to
communicate it all.

Not (yet) shown?  Spicy Sig format.  We've just carried all the
relevant data around in memory during this demo.  So: up next:
serializations for that.