friends should not let friends write concatenative formats without consistent delimiters
This format is *very nearly unparsable*, as it eschews *all of*:

- opening and closing delimiters
- length prefixes
- consistent separation delimiters

... *except* that last one is in force, by the skin of our teeth...

... in one of the two implementations I see working with these formats.

I'm going to continue implementing this, but I think it's remarkably
touchy.  This format should be reconsidered.

I've heard arguments that custom string-splitting formats like this can
described as "simple", and that "simple" prevents parser bugs and other
potential security risks.  However, I think finding behavioral
divergences already between the first and the second upstream libraries
to handle these formats... gives noteworthy empirical evidence that
this particular approach to "simple" is, in fact, not.

Simple is not its own identity value under the concatenation operator.