fourth commit on this: the constraint is dubious after all.
I thought we were clear for a moment, at that last commit; however,
I MADE AN UNRELATED MISTAKE IN THE TEST FIXTURES, which made the
parse of that fixture be rejected (as intended) but for an unrelated
reason (oh dear), and came to an incorrect conclusion from that.

There actually is NOT a constraint on Signed Note bodies that would
forbid them from containing the double-linebreak.  (So, what I thought
the first time, several commits back.)

I have now rewritten my considerations of the parse clarity
challenges here a third time.

(I would like to take this moment to also offer some rebuke towards
an idea that seems popular among some security-minded engineers,
being that giving specific parser error messages is always a bad idea.
Here, the lack of a clearer error message (athough admittedly
compounded with other misktakes of my own) almost lead me to make a
mistaken assumption that a property was checked... which is not.
While I caught it by reading my fixtures yet again, the
misunderstanding could've been intercepted much earlier by a more
meaningful error message from this parse code.  But I digress.)

So: overall:
we're back to "whether spicysig is parsable depends on who you ask".

Moving on...