a complete parser for spicysig.
Tests included; the fixture used is a real set of values produced
from demo_test (although I composed the strings manually, and the
realness isn't relevant here, because the MIP isn't being checked in
this test: only the parse).

It should be noted that I'm not at all sure this struct as written
shall be the right one if we want any reasonable API, even internally.
Storing the checkpoint as a Note value means we haven't yet actually
parsed the checkpoint itself (whoops); it's *also* a value that's
bizarrely hard to produce: it's only produced by parsing!  (While its
fields are public, it's only produced by `note.Open`; `note.Sign` goes
right to emitting bytes)... which leaves us in an odd position if we
want to do the work of signing at one moment and the serialization
at a later moment.