spicytool cli!  largely fresh take, using primitives now upstreamed into torchwood.
The new CLI so far supports the sign operations.
Verify coming up next.
Keygen probably shortly thereafter.

Most of the "spicy" package here will be dropped over the next few
commits.  The "glue" package can already be dropped (similar things
in torchwood 'tesserax' can replace it). Also the "signing" package
is already dropped, as its behaviors either moved upstream to features
in the torchwood library, or are moving closer to the CLI itself.

Things are getting cleaner and simpler with every iteration of this;
it's pretty lovely.