cli: verify, keygen.
Also: drop all remaining code for spicy format that now has upstream
implementations in torchwood that we're using instead.

I _believe_ this should end to end work... with a couple caveats:
mainly, I'm not quite sure what to do with the witness policy defaults
and the key management story for a signer.  (And, at this point that's
also become a blocker for testing anything all the way through.)

The default_policy file I've included here specifies several fairly
well-known witness services.  (It's cribbed from the age-keyserver
project.)  However, as far as those being used during signing,
where they're asked to perform new witnessing signatures: I'm not
actually sure what the criteria are for these services and if they'll
enthusiastically witness anyone who comes along and asks for it.

These policy files also can include an entry for the log itself and the
log's key, and while the age-keyserver and age-keylookup commands I'm
referencing heavily here can sensibly bake that key into their default,
here, where the intended user journey for spicytool is to have more
transparency logs started by a variety of operators, and where we want
the verify side of the tool to be able to work for any of those logs,
I'm less sure what a sensible set of values is to use here.